requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.22
sentence-transformers==3.0.1
chromadb==0.5.4
tiktoken==0.7.0
pydantic==2.9.2
python-dotenv==1.0.1
ragas==0.1.15
datasets==2.21.0
typer==0.12.5
pypdf==4.3.1

